Rename jarvis.py to core.py
